requirements.txt
absl-py==2.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
clarabel==0.6.0
colorama==0.4.6
contourpy==1.1.1
cvxpy==1.4.1
cycler==0.11.0
ecos==2.0.12
faiss-cpu==1.7.4
filelock==3.13.4
fonttools==4.42.1
fsspec==2024.3.1
grpcio==1.62.2
huggingface-hub==0.22.2
idna==3.4
imbalanced-learn==0.12.2
imblearn==0.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.7.3
mdurl==0.1.2
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.8.1.78
osqp==0.6.3
packaging==23.1
pandas==2.0.3
Pillow==9.4.0
protobuf==5.26.1
pybind11==2.11.1
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-metric-learning==0.9.99
pytz==2023.3.post1
PyYAML==6.0.1
qdldl==0.1.7.post0
requests==2.31.0
rich==13.3.5
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.10.1
scs==3.2.4.post1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
timm==0.9.16
tornado==6.3.3
tqdm==4.62.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
visdom==0.2.4
websocket-client==1.6.3
Werkzeug==3.0.2
